Removed the stupid db_int system from Config
Now Config only has one database which is string -> string. The getInt and
setInt functions are for convenience, which simply perform atoi and itoa
conversions on the fly.

Previously, db_int was intended as a cache to avoid doing int conversions on
every access, but this ended up bloating the code and required much more
std::map searches.